requirements.txt
googletrans==3.1.0a0
python-telegram-bot==13.2